+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * add `Controller.make_response` as an alias for `flask.make_response`
 * convert attributes on `Controller`, `Resource`, and `ModelResource` to be `class Meta` options
 * rename `_meta` to `Meta` per py-meta-utils v0.3
+* rename `ModelManager.find_all` to `ModelManager.all` and `ModelManager.find_by` to `ModelManager.filter_by` for consistency with the `Query` api
 
 ## 0.5.1 (2018/07/25)